+++ CHANGELOG.md
@@ -10,6 +10,9 @@ adheres to [Semantic Versioning][semver].
 
 ## [Unreleased]
 
+- Add a full dump of the payload when running with debugging.
+- Make the headers work correctly with older versions of curl.
+
 ## [v1.0.2] - 2021-10-04
 
 - Make it work correctly when you symlink to the `inform-slack` script
@@ -22,7 +25,7 @@ adheres to [Semantic Versioning][semver].
 
 - First actual release
 
-[Unreleased]: https://github.com/jasonk/inform-slack/compare/v1.0.2...HEAD
+[Unreleased]: https://github.com/jasonk/inform-slack/compare/v1.0.3...HEAD
 [v1.0.2]: https://github.com/jasonk/inform-slack/releases/tag/v1.0.2
 [v1.0.1]: https://github.com/jasonk/inform-slack/releases/tag/v1.0.1
 [v1.0.0]: https://github.com/jasonk/inform-slack/releases/tag/v1.0.0